sigrok plugin: Provide Averaging functionality
If the requested update rate is less than the refresh rate of the
device, we are losing precision and data. Instead of taking only
a sample every period - collate and average all the samples during
the period - and return the average.

This will ensure that the data precision taken by the samples are
not lost to the ether...

Signed-off-by: Kieran Bingham <kieran.bingham@st.com>