sigrok plugin: Provide consolidation configuration
If the requested update rate is less than the refresh rate of the
device, we are losing precision and data. Instead of taking only
a sample every period - we can consolidate this data into a single
sample point.

There are 4 methods of consolidation:

  Consolidation min|max|average|last

  min:     The smallest sample in the period is used
  max:     The largest sample point is taken
  average: An average of all samples over the period is provided.
  last:    The default method, where the last sample point is given

This will ensure that the data precision taken by the samples are
not lost, but can be used and reported in a user configurable way

To use this feature, provide the Consolidation option in the
<Device> configuration section as shown by the following example:

<Plugin sigrok>
  LogLevel 3
  <Device "inlet_current">
    Driver "agilent-dmm"
    MinimumInterval 10
    Consolidation average
    Conn "/dev/ttyUSB0"
  </Device>
</Plugin>

Signed-off-by: Kieran Bingham <kieran.bingham@st.com>